+++ Control/Joint/Abilities/Adaptable.hs
@@ -1,65 +1,65 @@
-module Control.Joint.Abilities.Liftable where
+module Control.Joint.Abilities.Adaptable where
 
 import Control.Joint.Core (type (~>))
 import Control.Joint.Abilities.Interpreted (Interpreted (run))
 import Control.Joint.Abilities.Transformer (Transformer (Schema, build, embed), (:>) (T, trans))
 
-class Liftable (eff :: * -> *) (schema :: * -> *) where
-	{-# MINIMAL lift #-}
-	lift :: eff ~> schema
+class Adaptable (eff :: * -> *) (schema :: * -> *) where
+	{-# MINIMAL adapt #-}
+	adapt :: eff ~> schema
 
 type Embedding t u = (Transformer t, Functor u)
 type Building t u = (Transformer t, Applicative u)
 
-instance Liftable t t where
-	lift = id
+instance Adaptable t t where
+	adapt = id
 
-instance Embedding t u => Liftable u (t :> u) where
-	lift = embed
+instance Embedding t u => Adaptable u (t :> u) where
+	adapt = embed
 
-instance Building t u => Liftable t (t :> u) where
-	lift = build
+instance Building t u => Adaptable t (t :> u) where
+	adapt = build
 
 instance
 	( Embedding t (Schema u v)
 	, Building u v
-	) => Liftable u (t :> u :> v) where
-	lift = embed . build
+	) => Adaptable u (t :> u :> v) where
+	adapt = embed . build
 
 instance
 	( Embedding t (Schema u v)
 	, Embedding u v
-	) => Liftable v (t :> u :> v) where
-	lift = embed . embed
+	) => Adaptable v (t :> u :> v) where
+	adapt = embed . embed
 
 instance
 	( Embedding t (Schema u (v :> w))
 	, Embedding u (Schema v w)
 	, Building v w
-	) => Liftable v (t :> u :> v :> w) where
-	lift = embed . embed . build
+	) => Adaptable v (t :> u :> v :> w) where
+	adapt = embed . embed . build
 
 instance
 	( Embedding t (Schema u v)
 	, Embedding t (Schema u (v :> w))
 	, Embedding u (Schema v w)
 	, Embedding v w
-	) => Liftable w (t :> u :> v :> w) where
-	lift = embed . embed . embed
+	) => Adaptable w (t :> u :> v :> w) where
+	adapt = embed . embed . embed
 
 instance (Embedding t (Schema u (v :> w :> x))
 	, Embedding u (Schema v (w :> x))
 	, Embedding v (Schema w x)
 	, Embedding w x
-	) => Liftable x (t :> u :> v :> w :> x) where
-	lift = embed . embed . embed . embed
+	) => Adaptable x (t :> u :> v :> w :> x) where
+	adapt = embed . embed . embed . embed
 
 instance (Embedding t (Schema u (v :> w :> x))
 	, Embedding u (Schema v (w :> x))
 	, Embedding v (Schema w x)
 	, Building w x
-	) => Liftable w (t :> u :> v :> w :> x) where
-	lift = embed . embed . embed . build
+	) => Adaptable w (t :> u :> v :> w :> x) where
+	adapt = embed . embed . embed . build
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y))
@@ -67,8 +67,8 @@ instance
 	, Embedding v (Schema w (x :> y))
 	, Embedding w (Schema x y)
 	, Embedding x y
-	) => Liftable y (t :> u :> v :> w :> x :> y) where
-	lift = embed . embed . embed . embed . embed
+	) => Adaptable y (t :> u :> v :> w :> x :> y) where
+	adapt = embed . embed . embed . embed . embed
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y))
@@ -76,8 +76,8 @@ instance
 	, Embedding v (Schema w (x :> y))
 	, Embedding w (Schema x y)
 	, Building x y
-	) => Liftable x (t :> u :> v :> w :> x :> y) where
-	lift = embed . embed . embed . embed . build
+	) => Adaptable x (t :> u :> v :> w :> x :> y) where
+	adapt = embed . embed . embed . embed . build
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z))
@@ -86,8 +86,8 @@ instance
 	, Embedding w (Schema x (y :> z))
 	, Embedding x (Schema y z)
 	, Embedding y z
-	) => Liftable z (t :> u :> v :> w :> x :> y :> z) where
-	lift = embed . embed . embed . embed . embed . embed
+	) => Adaptable z (t :> u :> v :> w :> x :> y :> z) where
+	adapt = embed . embed . embed . embed . embed . embed
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z))
@@ -96,8 +96,8 @@ instance
 	, Embedding w (Schema x (y :> z))
 	, Embedding x (Schema y z)
 	, Building y z
-	) => Liftable y (t :> u :> v :> w :> x :> y :> z) where
-	lift = embed . embed . embed . embed . embed . build
+	) => Adaptable y (t :> u :> v :> w :> x :> y :> z) where
+	adapt = embed . embed . embed . embed . embed . build
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z :> f))
@@ -107,8 +107,8 @@ instance
 	, Embedding x (Schema y (z :> f))
 	, Embedding y (Schema z f)
 	, Embedding z f
-	) => Liftable f (t :> u :> v :> w :> x :> y :> z :> f) where
-	lift = embed . embed . embed . embed . embed . embed . embed
+	) => Adaptable f (t :> u :> v :> w :> x :> y :> z :> f) where
+	adapt = embed . embed . embed . embed . embed . embed . embed
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z :> f))
@@ -118,8 +118,8 @@ instance
 	, Embedding x (Schema y (z :> f))
 	, Embedding y (Schema z f)
 	, Building z f
-	) => Liftable z (t :> u :> v :> w :> x :> y :> z :> f) where
-	lift = embed . embed . embed . embed . embed . embed . build
+	) => Adaptable z (t :> u :> v :> w :> x :> y :> z :> f) where
+	adapt = embed . embed . embed . embed . embed . embed . build
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z :> f :> h))
@@ -130,8 +130,8 @@ instance
 	, Embedding y (Schema z (f :> h))
 	, Embedding z (Schema f h)
 	, Embedding f h
-	) => Liftable h (t :> u :> v :> w :> x :> y :> z :> f :> h) where
-	lift = embed . embed . embed . embed . embed . embed . embed . embed
+	) => Adaptable h (t :> u :> v :> w :> x :> y :> z :> f :> h) where
+	adapt = embed . embed . embed . embed . embed . embed . embed . embed
 
 instance
 	( Embedding t (Schema u (v :> w :> x :> y :> z :> f :> h))
@@ -142,5 +142,5 @@ instance
 	, Embedding y (Schema z (f :> h))
 	, Embedding z (Schema f h)
 	, Building f h
-	) => Liftable f (t :> u :> v :> w :> x :> y :> z :> f :> h) where
-	lift = embed . embed . embed . embed . embed . embed . embed . build
+	) => Adaptable f (t :> u :> v :> w :> x :> y :> z :> f :> h) where
+	adapt = embed . embed . embed . embed . embed . embed . embed . build