+++ Control/Joint/Effects/Writer.hs
@@ -3,7 +3,7 @@ module Control.Joint.Effects.Writer where
 import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
 import Control.Joint.Abilities.Transformer (Transformer (Schema, embed, build, unite), (:>) (T))
-import Control.Joint.Abilities.Liftable (Liftable (lift))
+import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (UT (UT))
 
 newtype Writer e a = Writer (e, a)
@@ -40,7 +40,7 @@ instance (Monoid e, Applicative u) => Applicative (UT ((,) e) u) where
 instance (Monoid e, Applicative u, Monad u) => Monad (UT ((,) e) u) where
 	UT x >>= f = UT $ x >>= \(acc, v) -> (\(acc', y) -> (acc <> acc', y)) <$> run (f v)
 
-type Accumulated e t = Liftable (Writer e) t
+type Accumulated e t = Adaptable (Writer e) t
 
 add :: Accumulated e t => e -> t ()
-add s = lift $ Writer (s, ())
+add s = adapt $ Writer (s, ())